% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRRR-package.R
\docType{package}
\name{RRRR-package}
\alias{RRRR-package}
\title{Online Robust Reduced-Rank Regression Estimation}
\description{
Methods for estimating online Robust Reduced-Rank Regression.
}
\author{
Yangzhuoran Yang. \email{yangyangzhuoran@gmail.com}

Ziping Zhao. \email{zhaoziping@shanghaitech.edu.cn}
}
\keyword{package}
