% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update.RRRR}
\alias{update.RRRR}
\title{Update the RRRR/ORRRR type model with addition data}
\usage{
\method{update}{RRRR}(
  object,
  newy,
  newx,
  newz = NULL,
  addon = object$spec$addon,
  method = object$method,
  SAAmethod = object$SAAmethod,
  ...,
  ProgressBar = requireNamespace("lazybar")
)
}
\arguments{
\item{object}{A model with class \code{RRRR}(\code{ORRRR})}

\item{newy}{Matrix of dimension N*P, the new data y. The matrix for the response variables. See \code{Detail}.}

\item{newx}{Matrix of dimension N*Q, the new data x. The matrix for the explanatory variables to be projected. See \code{Detail}.}

\item{newz}{Matrix of dimension N*R, the new data z. The matrix for the explanatory variables not to be projected. See \code{Detail}.}

\item{addon}{Integer. The number of data points to be added in the algorithm in each iteration after the first.}

\item{method}{Character. The estimation method. Either "SMM" or "SAA". See \code{Description}.}

\item{SAAmethod}{Character. The sub solver used in each iteration when the \code{methid} is chosen to be "SAA". See \code{Detail}.}

\item{...}{Additional arguments to function
\describe{
\item{\code{optim}}{when the \code{method} is "SAA" and the \code{SAAmethod} is "optim"}
\item{\code{RRRR}}{when the \code{method} is "SAA" and the \code{SAAmethod} is "MM"}
}}

\item{ProgressBar}{Logical. Indicating if a progress bar is shown during the estimation process.
The progress bar requires package \code{lazybar} to work.}
}
\value{
A list of the estimated parameters of class \code{ORRRR}.
\describe{
\item{method}{The estimation method being used}
\item{SAAmethod}{If SAA is the major estimation method, what is the sub solver in each iteration.}
\item{spec}{The input specifications. \eqn{N} is the sample size.}
\item{history}{The path of all the parameters during optimization and the path of the objective value.}
\item{mu}{The estimated constant vector. Can be \code{NULL}.}
\item{A}{The estimated exposure matrix.}
\item{B}{The estimated factor matrix.}
\item{D}{The estimated coefficient matrix of \code{z}.}
\item{Sigma}{The estimated covariance matrix of the innovation distribution.}
\item{obj}{The final objective value.}
\item{data}{The data used in estimation.}
}
}
\description{
\code{update.RRRR} will update online robust reduced-rank regression model with class \code{RRRR}(\code{ORRRR}) using newly added data
to achieve online estimation.
Estimation methods:
\describe{
\item{SMM}{Stochastic Majorisation-Minimisation}
\item{SAA}{Sample Average Approximation}
}
}
\details{
The formulation of the reduced-rank regression is as follow:
\deqn{y = \mu +AB'  x + D z+innov,}
where for each realization \eqn{y} is a vector of dimension \eqn{P} for the \eqn{P} response variables,
\eqn{x} is a vector of dimension \eqn{Q} for the \eqn{Q} explanatory variables that will be projected to
reduce the rank,
\eqn{z} is a vector of dimension \eqn{R} for the \eqn{R} explanatory variables
that will not be projected,
\eqn{\mu} is the constant vector of dimension \eqn{P},
\eqn{innov} is the innovation vector of dimension \eqn{P},
\eqn{D} is a coefficient matrix for \eqn{z} with dimension \eqn{P*R},
\eqn{A} is the so called exposure matrix with dimension \eqn{P*r}, and
\eqn{B} is the so called factor matrix with dimension \eqn{Q*r}.
The matrix resulted from \eqn{AB'} will be a reduced rank coefficient matrix with rank of \eqn{r}.
The function estimates parameters \eqn{\mu}, \eqn{A}, \eqn{B}, \eqn{D}, and \eqn{Sigma}, the covariance matrix of
the innovation's distribution.

See \code{?ORRRR} for details about the estimation methods.
}
\examples{
\donttest{
set.seed(2222)
data <- RRR_sim()
newdata <- RRR_sim(A = data$spec$A,
                   B = data$spec$B,
                   D = data$spec$D)
res <- ORRRR(y=data$y, x=data$x, z = data$z)
res <- update(res, newy=newdata$y, newx=newdata$x, newz=newdata$z)
res
}
}
\seealso{
\code{ORRRR}, \code{RRRR}, \code{RRR}
}
\author{
Yangzhuoran Yang
}
